# application/services/custom_services/nutrition/src/models/Mealboxmanager.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const MealboxmanagerSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   name: String,
   kcal: Number,
   carbohydrates: Number,
   protein: Number,
   portion: Number
})

const MealboxmanagerModel = mongoose.model('Mealboxmanager', MealboxmanagerSchema, 'Mealboxmanager');
export default MealboxmanagerModel;