# application/services/custom_services/nutrition/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: healthmonkstart62 open api specification
   description: nutrition API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8011
     description: nutrition
tags:
- name: Mealboxmanager
paths:
   "/Mealboxmanager/{id}":
      delete:
         tags:
         -  Mealboxmanager
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  Mealboxmanager
         description: GpGetNounById
         parameters:
         -  in: "path"
            description: GpGetNounById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/Mealboxmanager/get/search":
      get:
         tags:
         -  Mealboxmanager
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/Mealboxmanager/get/update":
      put:
         tags:
         -  Mealboxmanager
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/Mealboxmanager"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/Mealboxmanager":
      put:
         tags:
         -  Mealboxmanager
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/Mealboxmanager"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  Mealboxmanager
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  Mealboxmanager
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/Mealboxmanager"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      Mealboxmanager:
         type: object
         properties:
            name:
               type: "string"
               description: Description
            kcal:
               type: "number"
               description: Description
            carbohydrates:
               type: "number"
               description: Description
            protein:
               type: "number"
               description: Description
            portion:
               type: "number"
               description: Description